Narrow the URL helper: rename navigation -> browserLocation, drop <Navigate>
The helper is not a router; it is same-page URL state plus real browser navigation
for path changes. Rename `navigation.tsx` -> `browserLocation.tsx` so the name
reflects that boundary, and update all importers.

Remove the `<Navigate>` declarative-redirect export and its two call sites
(EipPage, DevnetSpecPage). Client-side route normalization belongs to Astro now:
getStaticPaths emits only valid detail pages, configured redirects handle aliases,
and unknown paths fall through to Astro's 404 — so those fallbacks were
unreachable. Also drops DevnetSpecPage's now-dead `loading` guard (the snapshot
hook resolves synchronously).
