+++ src/components/UpgradesIndexPage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Link } from './navigation';
+import { Link } from './browserLocation';
 import { networkUpgrades, NetworkUpgrade } from '../data/upgrades';
 import { parseShortDate } from './schedule/forkDateCalculator';
 import UpgradeCard from './ui/UpgradeCard';
