+++ src/components/calls-index/CallsIndexTimeline.tsx
@@ -1,4 +1,4 @@
-import { Link } from '../navigation';
+import { Link } from '../browserLocation';
 import { callTypeNames, isOneOffCall, type Call, type CallType } from '../../data/calls';
 import { hasUpcomingWatchPage } from '../../domain/calls/upcomingCalls';
 import { type TimelineEvent } from '../../data/events';
