+++ src/components/eip/EipNotice.tsx
@@ -1,5 +1,5 @@
 import type { ReactNode } from 'react';
-import { Link } from '../navigation';
+import { Link } from '../browserLocation';
 import { ForkRelationship } from '../../types';
 import { parseMarkdownLinks } from '../../utils';
 import { formatCallReference } from '../../domain/calls/callReference';
